+++ src/Products/PloneMeeting/migrations/migrate_to_4111.py
@@ -11,12 +11,15 @@ class Migrate_To_4111(Migrator):
         # make sure portal_type ConfigurablePODTemplate is
         # reapplied to add store_as_annex_empty_file field
         self.ps.runImportStepFromProfile('profile-Products.PloneMeeting:default', 'typeinfo')
+        # update holidays
+        self.updateHolidays()
 
 
 def migrate(context):
     '''This migration function will:
 
-       1) Re-apply 'typeinfo' step to get ConfigurablePODTemplate.store_as_annex_empty_file.
+       1) Re-apply 'typeinfo' step to get ConfigurablePODTemplate.store_as_annex_empty_file;
+       2) Update holidays to take 2021 into account.
     '''
     migrator = Migrate_To_4111(context)
     migrator.run()